fix kill for wait

+++ scripts/fivestar/daemon.sh
@@ -32,5 +32,7 @@ else
 fi
 
 wait "$pid"
-kill $pid
+if [ ! -z "$pid"]; then
+    kill $pid
+fi
