+++ scripts/fivestar/worker.sh
@@ -95,4 +95,6 @@ else
 fi
 
 wait "$pid"
-kill $pid
+if [ ! -z "$pid"]; then
+    kill $pid
+fi
